feat: create unit tests for visualiser component
